sample-backend/requirements.txt
flask
opentelemetry-distro
opentelemetry-exporter-otlp
python-dotenv
flask_cors